int to size_t and added comments

+++ project/include/utils.h
@@ -1,7 +1,8 @@
+#include <stdlib.h>
 #ifndef PROJECT_INCLUDE_UTILS_H_
 #define PROJECT_INCLUDE_UTILS_H_
 
-int timer_from(unsigned int from);
+size_t timer_from(unsigned int from);
 long long int custom_pow(int base, int pow);
   // TODO(Ivan Biryukov): Stuff. Implement `power of` function
